# api/openapi.yaml
openapi: 3.1.0
info:
  title: OCS Testbench API
  version: 0.2.0
  description: |
    REST API for the OCS Testbench UI and external clients.

    This specification is the single source of truth for the HTTP contract
    between the Go backend (`internal/api`) and the React frontend (`web/`).
    Generated TypeScript types live at `web/src/api/schema.d.ts` — regenerate
    after any change to this file.

    See `docs/ARCHITECTURE.md` for behavioural semantics (state machines,
    invariants, variable resolution rules). This document is the wire
    contract; it does not restate the semantics.

    Real-time server → client events (peer status changes, execution
    progress, pause / resume transitions) are delivered via the SSE
    endpoint at `GET /events`.
servers:
  - url: /api/v1
    description: Current server (relative to origin)
...
paths:
...
  /peers:
    get:
      tags: [Peers]
      operationId: listPeers
      summary: List configured peers with current connection status
      responses:
        '200':
          description: Peer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Peer'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags: [Peers]
      operationId: createPeer
      summary: Create a new peer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PeerInput'
      responses:
        '201':
          description: Peer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Peer'
        '422':
          $ref: '#/components/responses/ValidationProblem'
        default:
          $ref: '#/components/responses/Problem'
...
components:
...
  responses:
    Problem:
      description: Error response (RFC 7807)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationProblem:
      description: |
        Validation error (RFC 7807). Carries field-level error messages
        keyed by JSON Pointer references into the request body.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    Problem:
      type: object
      description: RFC 7807 problem details
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: object
          description: |
            Optional field-level validation errors. Keys are JSON Pointer
            references into the request body (e.g. `/name`, `/services/0/ratingGroup`);
            values are arrays of human-readable error messages. Present on
            `422` responses and ignorable otherwise.
          additionalProperties:
            type: array
            items:
              type: string
            minItems: 1

    # ─────────────── Pagination ───────────────
...
    PeerStatus:
      type: string
      description: |
        Current runtime state of a peer.

        - `stopped`       — Administratively down; no reconnect attempts.
        - `disconnected`  — Supervision on, transport down; auto-redials.
        - `connecting`    — CER/CEA exchange in progress.
        - `connected`     — Session established.
        - `disconnecting` — Graceful teardown in progress.
        - `restarting`    — Admin-initiated stop+start for a config change.
        - `error`         — Supervision on, persistent failure, exp. backoff.

        `stopped` vs `disconnected`: admin intent vs transport health.
      enum:
        - stopped
        - disconnected
        - connecting
        - connected
        - disconnecting
        - restarting
        - error
    PeerTransport:
      type: string
      description: Transport protocol used for the Diameter connection.
      enum: [TCP, TLS]
    Peer:
      type: object
      required:
        [id, name, host, port, originHost, originRealm, transport, watchdogIntervalSeconds, autoConnect, status]
      properties:
        id:
          type: string
        name:
          type: string
        host:
          type: string
          examples: ['10.0.1.5']
        port:
          type: integer
          minimum: 1
          maximum: 65535
          examples: [3868]
        originHost:
          type: string
          examples: ['ctf-01.test.local']
        originRealm:
          type: string
          examples: ['test.local']
        transport:
          $ref: '#/components/schemas/PeerTransport'
        watchdogIntervalSeconds:
          type: integer
          minimum: 5
          maximum: 3600
          default: 30
        autoConnect:
          type: boolean
          description: |
            Pure server-boot flag. When the server starts, peers with
            `autoConnect: true` are started automatically; peers with
            `autoConnect: false` remain `stopped` until operator action.
          default: true
        status:
          $ref: '#/components/schemas/PeerStatus'
        statusDetail:
          type: string
          description: Optional human-readable status detail
        lastChangeAt:
          type: string
          format: date-time
    PeerInput:
      type: object
      description: |
        Writable subset of `Peer`. Used for both create and replace.
        Live-state fields are computed server-side and rejected on write.
      required: [name, host, port, originHost, originRealm, transport, watchdogIntervalSeconds, autoConnect]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        host:
          type: string
          minLength: 1
        port:
          type: integer
          minimum: 1
          maximum: 65535
        originHost:
          type: string
          minLength: 1
        originRealm:
          type: string
          minLength: 1
        transport:
          $ref: '#/components/schemas/PeerTransport'
        watchdogIntervalSeconds:
          type: integer
          minimum: 5
          maximum: 3600
          default: 30
        autoConnect:
          type: boolean
          default: true
...